requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0 